Swing.c: Swing_ApplyForgiveness 97.4 -> 99.5 (s8 nAttr)

+++ src/Swing.c
@@ -292,7 +292,7 @@ void Swing_ApplyForgiveness(int nPlayer) {
     {
     f32  fError = gPlayers[nPlayer].swing.fMishitAngle;
     int  nRowScale, nRowThresh;
-    int  nAttr;
+    s8   nAttr;
     f32  fThresh, fScale;
 
     if (gPlayers[nPlayer].ball.nLie == 6 || gPlayers[nPlayer].ball.nLie == 7 ||
